requirements.txt
Flask==3.0.0
httpx==0.28.1
groq==0.4.2
plotly==5.18.0
python-dotenv==1.0.0